pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentgateway"
version = "0.1.0"
description = "MCP-compatible integration gateway for LLM agents with evaluation, observability, and multi-tenant routing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "anthropic>=0.28.0",
    "openai>=1.30.0",
    "langgraph>=0.1.0",
    "langchain>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",
    "structlog>=24.0.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
